rename restrict operator to filter (#16374)
rename restrict operator to filter

Approved by: @ouyuanning, @m-schen, @XuPeng-SH
