+++ pkg/sql/colexec/filter/filter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package restrict
+package filter
 
 import (
 	"bytes"
@@ -29,7 +29,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 )
 
-const argName = "restrict"
+const argName = "filter"
 
 func (arg *Argument) String(buf *bytes.Buffer) {
 	buf.WriteString(argName)
