+++ pkg/sql/colexec/filter/types.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package restrict
+package filter
 
 import (
 	"github.com/matrixorigin/matrixone/pkg/common/reuse"
