+++ pkg/vm/overload.go
@@ -33,7 +33,7 @@ package vm
 // 	Output:      output.String,
 // 	Offset:      offset.String,
 // 	Product:     product.String,
-// 	Restrict:    restrict.String,
+// 	Filter:    restrict.String,
 // 	Dispatch:    dispatch.String,
 // 	Connector:   connector.String,
 // 	Projection:  projection.String,
@@ -103,7 +103,7 @@ package vm
 // 	Output:      output.Prepare,
 // 	Offset:      offset.Prepare,
 // 	Product:     product.Prepare,
-// 	Restrict:    restrict.Prepare,
+// 	Filter:    restrict.Prepare,
 // 	Dispatch:    dispatch.Prepare,
 // 	Connector:   connector.Prepare,
 // 	Projection:  projection.Prepare,
@@ -173,7 +173,7 @@ package vm
 // 	Output:      output.Call,
 // 	Offset:      offset.Call,
 // 	Product:     product.Call,
-// 	Restrict:    restrict.Call,
+// 	Filter:    restrict.Call,
 // 	Dispatch:    dispatch.Call,
 // 	Connector:   connector.Call,
 // 	Projection:  projection.Call,
